Adjustments, commit No.9, 2019/5/23 20:53.

+++ Server/project/check/views.py
@@ -25,26 +25,31 @@ def upload(request):
     uploadFiles = request.FILES.getlist('file')
 
     if CaseFiles.objects.filter(SN=SN):
-        fs = FileSystemStorage()
-        path = os.path.abspath('.') + "/uploads"
-        destination = os.path.abspath('.') + "/check/casefiles/case" + SN
+        if Case.objects.get(SN=SN).checked == 0:
+            fs = FileSystemStorage()
+            path = os.path.abspath('.') + "/uploads"
+            destination = os.path.abspath('.') + "/check/casefiles/case" + SN
 
-        for f in uploadFiles:
-            if f.name.endswith('.html'):
-                fs.save('result'+SN+'.html', f)
-            else:
-                fs.save(f.name, f)
+            for f in uploadFiles:
+                if f.name.endswith('.html'):
+                    fs.save('result'+SN+'.html', f)
+                else:
+                    fs.save(f.name, f)
 
-        for f in os.listdir(destination):
-            os.remove(f)
+            for f in os.listdir(destination):
+                os.remove(f)
 
-        for f in os.listdir(path):
-            shutil.move(path + "/" + f, destination)
+            for f in os.listdir(path):
+                shutil.move(path + "/" + f, destination)
 
-        Case.objects.filter(SN=SN).update(checked=1)
+            Case.objects.filter(SN=SN).update(checked=1)
 
-        return HttpResponse(json.dumps({'statusCode': 'success'}),
-                            content_type="application/json")
+            return HttpResponse(json.dumps({'statusCode': 'success'}),
+                                content_type="application/json")
+
+        else:
+            return HttpResponse(json.dumps({'statusCode': 'exist'}),
+                                content_type="application/json")
     else:
         return HttpResponse(json.dumps({'statusCode': 'failed'}),
                             content_type="application/json")
@@ -64,13 +69,42 @@ def result(request):
 
 @group_required('Volunteer', 'Engineer')
 def showUnassignedCases(request):
-    data = Case.objects.all()
+    data = Case.objects.filter(assign='0')
     response = []
     for d in data:
-        if d.assign == '0':
-            response.append(d.name + " " + d.SN)
-    for r in response:
-        print(r)
+        response.append(d.SN + " " + d.name)
+    return HttpResponse(response)
+
+
+@group_required('Volunteer', 'Engineer')
+def showCheckedCases(request):
+    data = Case.objects.filter(volunteer=request.user.username, checked='1')
+    response = []
+    for d in data:
+        response.append(d.SN + " " + d.name)
+    return HttpResponse(response)
+
+
+@group_required('Volunteer', 'Engineer')
+def showMyCases(request):
+    data = Case.objects.filter(username=request.user.username, checked='0')
+    response = []
+    for d in data:
+        response.append(d.SN + " " + d.name)
+    return HttpResponse(response)
+
+
+@group_required('Volunteer', 'Engineer')
+def showDetail(request):
+    case = Case.objects.get(name=request.GET.get('name'),
+                            address=request.GET.get('address'))
+    response = []
+    response.append(case.name)
+    response.append(case.buildingType)
+    response.append(case.address)
+    response.append(case.phone)
+    response.append(case.applyDate)
+    return HttpResponse(response)
 
 
 @group_required('Volunteer', 'Engineer')
